refactor(backoffice): remove old Bash runtime architecture
