+++ apps/backoffice/app/fragno/automation/engine/bash.test.ts
@@ -1,64 +1,17 @@
 import { describe, expect, it } from "vitest";
 
-import { Bash, InMemoryFs } from "just-bash";
+import { InMemoryFs } from "just-bash";
 
 import { MasterFileSystem } from "@/files/master-file-system";
 import { normalizeMountedFileSystem } from "@/files/mounted-file-system";
+import { createBashHost, executeBashAutomation } from "@/fragno/runtime-tools/bash-host";
 
-import { createBashHost, executeBashAutomation } from "../../bash-runtime/bash-host";
-import { createAutomationCommands } from "../commands/bash-adapter";
-import {
-  AUTOMATIONS_COMMAND_SPEC_LIST,
-  EVENT_COMMAND_SPEC_LIST,
-  OTP_COMMAND_SPEC_LIST,
-} from "../commands/registry";
-import type {
-  AutomationCommandContext,
-  AutomationsCommandHandlers,
-  BashAutomationCommandResult,
-  EventCommandHandlers,
-  OtpCommandHandlers,
-} from "../commands/types";
 import type { AutomationEvent } from "../contracts";
 import {
   createAutomationExecutionContext as createRuntimeAutomationExecutionContext,
   createAutomationRuntime as createStorageBackedAutomationRuntime,
   type AutomationRuntime,
 } from "./runtime";
-
-type TestContext = {
-  invocation: string[];
-} & AutomationCommandContext;
-
-type TestAutomationCommandHandlers = AutomationsCommandHandlers<TestContext> &
-  OtpCommandHandlers<TestContext> &
-  EventCommandHandlers<TestContext>;
-
-const TEST_COMMAND_SPEC_LIST = [
-  ...AUTOMATIONS_COMMAND_SPEC_LIST,
-  ...OTP_COMMAND_SPEC_LIST,
-  ...EVENT_COMMAND_SPEC_LIST,
-] as const;
-
-const createTestCommandContext = (overrides: Partial<TestContext> = {}): TestContext => ({
-  invocation: [],
-  event: {
-    id: "test-event",
-    source: "telegram",
-    eventType: "message.received",
-    occurredAt: "2026-01-01T00:00:00.000Z",
-    payload: {},
-  },
-  binding: {
-    source: "telegram",
-    eventType: "message.received",
-    scriptId: "script-ctx",
-  },
-  orgId: "org-ctx",
-  idempotencyKey: "idempotency-ctx",
-  bashEnv: {},
-  ...overrides,
-});
 
 const runtime: AutomationRuntime = {
   lookupBinding: async () => null,
@@ -120,575 +73,6 @@ const createTestAutomationExecutionContext = ({
   });
 
 describe("bash command runner", () => {
-  it("runs cli-style automation commands through just-bash", async () => {
-    const commandCallsResult: BashAutomationCommandResult[] = [];
-    const context = createTestCommandContext();
-
-    const handlers: TestAutomationCommandHandlers = {
-      "otp.identity.create-claim": async (command) => {
-        context.invocation.push("create");
-        return {
-          data: {
-            url: `https://example.com/${command.args.externalActorId}`,
-            externalId: command.args.externalActorId,
-            code: "123456",
-          },
-        };
-      },
-      "automations.identity.lookup-binding": async (command) => {
-        context.invocation.push("lookup");
-        if (command.args.key === "missing") {
-          return { exitCode: 1 };
-        }
-        return {
-          data: {
-            value: `user-for-${command.args.key}`,
-          },
-        };
-      },
-      "automations.identity.bind-actor": async (command) => {
-        context.invocation.push("bind");
-        return {
-          data: {
-            value: command.args.value,
-            key: command.args.key,
-          },
-        };
-      },
-      "event.emit": async (command) => {
-        context.invocation.push("emit");
-        return {
-          data: {
-            eventType: command.args.eventType,
-            source: command.args.source ?? "otp",
-          },
-        };
-      },
-      "scripts.run": async () => {
-        throw new Error("not available in test");
-      },
-    };
-
-    const fs = new InMemoryFs();
-    const bash = new Bash({
-      fs,
-      customCommands: createAutomationCommands(
-        TEST_COMMAND_SPEC_LIST,
-        handlers,
-        context,
-        commandCallsResult,
-      ),
-    });
-
-    const result = await bash.exec(
-      'claim_url="$(otp.identity.create-claim --source telegram --external-actor-id actor_1 --print url)"\n' +
-        'linked_user="$(automations.identity.lookup-binding --source telegram --key actor_1 --print value)"\n' +
-        'missing_user="$(automations.identity.lookup-binding --source telegram --key missing --print value || true)"\n' +
-        'automations.identity.bind-actor --source telegram --key actor_1 --value "$linked_user" --format json\n' +
-        "event.emit --event-type identity.binding.completed --source otp --format json\n" +
-        'echo "$claim_url|$linked_user|$missing_user"',
-    );
-
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout?.trim().split("\n").at(-1)).toBe(
-      "https://example.com/actor_1|user-for-actor_1|",
-    );
-    expect(context.invocation).toEqual(["create", "lookup", "lookup", "bind", "emit"]);
-    expect(commandCallsResult).toEqual([
-      {
-        command: "otp.identity.create-claim",
-        output: "https://example.com/actor_1",
-        exitCode: 0,
-      },
-      {
-        command: "automations.identity.lookup-binding",
-        output: "user-for-actor_1",
-        exitCode: 0,
-      },
-      {
-        command: "automations.identity.lookup-binding",
-        output: "",
-        exitCode: 1,
-      },
-      {
-        command: "automations.identity.bind-actor",
-        output: '{"value":"user-for-actor_1","key":"actor_1"}',
-        exitCode: 0,
-      },
-      {
-        command: "event.emit",
-        output: '{"eventType":"identity.binding.completed","source":"otp"}',
-        exitCode: 0,
-      },
-    ]);
-  });
-
-  it("shows command-line help for automation commands", async () => {
-    const commandCallsResult: BashAutomationCommandResult[] = [];
-    const context = createTestCommandContext();
-    const handlers: TestAutomationCommandHandlers = {
-      "otp.identity.create-claim": async () => {
-        return {
-          data: {
-            url: "https://example.com/help",
-            externalId: "external-id",
-            code: "help",
-          },
-        };
-      },
-      "automations.identity.lookup-binding": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-        },
-      }),
-      "automations.identity.bind-actor": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-          key: "external-id",
-          status: "linked",
-        },
-      }),
-      "event.emit": async () => ({
-        data: {
-          accepted: true,
-          eventId: "help-event",
-          source: "telegram",
-          eventType: "help-event",
-        },
-      }),
-      "scripts.run": async () => {
-        throw new Error("not available in test");
-      },
-    };
-
-    const bash = new Bash({
-      customCommands: createAutomationCommands(
-        TEST_COMMAND_SPEC_LIST,
-        handlers,
-        context,
-        commandCallsResult,
-      ),
-    });
-
-    const result = await bash.exec("otp.identity.create-claim --help");
-
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toContain("otp.identity.create-claim");
-    expect(result.stdout).toContain("--source");
-    expect(result.stdout).toContain("Usage:");
-    expect(result.stdout).toContain("--help");
-    expect(result.stdout).toContain("--print <selector>");
-    expect(result.stdout).toContain("--format <format>");
-    expect(context.invocation).toEqual([]);
-    expect(commandCallsResult).toEqual([
-      {
-        command: "otp.identity.create-claim",
-        output: expect.any(String),
-        exitCode: 0,
-      },
-    ]);
-  });
-
-  it("shows command-line help for all built-in automation commands", async () => {
-    const commandCallsResult: BashAutomationCommandResult[] = [];
-    const context = createTestCommandContext();
-    const handlers: TestAutomationCommandHandlers = {
-      "otp.identity.create-claim": async () => ({
-        data: { url: "https://example.com", externalId: "external-id", code: "123456" },
-      }),
-      "automations.identity.lookup-binding": async () => ({
-        data: { value: "help-user", source: "telegram" },
-      }),
-      "automations.identity.bind-actor": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-          key: "external-id",
-          status: "linked",
-        },
-      }),
-      "event.emit": async () => ({
-        data: {
-          accepted: true,
-          eventId: "help-event",
-          source: "telegram",
-          eventType: "help-event",
-        },
-      }),
-      "scripts.run": async () => {
-        throw new Error("not available in test");
-      },
-    };
-
-    const bash = new Bash({
-      customCommands: createAutomationCommands(
-        TEST_COMMAND_SPEC_LIST,
-        handlers,
-        context,
-        commandCallsResult,
-      ),
-    });
-
-    const helpResults = [
-      await bash.exec("otp.identity.create-claim --help"),
-      await bash.exec("automations.identity.lookup-binding --help"),
-      await bash.exec("automations.identity.bind-actor --help"),
-      await bash.exec("event.emit --help"),
-    ];
-
-    expect(helpResults.every((result) => result.exitCode === 0)).toBe(true);
-    for (const result of helpResults) {
-      expect(result.stdout).toContain("Usage:");
-      expect(result.stdout).toContain("--help");
-      expect(result.stdout).toContain("--print <selector>");
-      expect(result.stdout).toContain("--format <format>");
-    }
-    expect(context.invocation).toEqual([]);
-    expect(commandCallsResult).toEqual([
-      {
-        command: "otp.identity.create-claim",
-        output: expect.stringContaining("otp.identity.create-claim"),
-        exitCode: 0,
-      },
-      {
-        command: "automations.identity.lookup-binding",
-        output: expect.stringContaining("automations.identity.lookup-binding"),
-        exitCode: 0,
-      },
-      {
-        command: "automations.identity.bind-actor",
-        output: expect.stringContaining("automations.identity.bind-actor"),
-        exitCode: 0,
-      },
-      {
-        command: "event.emit",
-        output: expect.stringContaining("event.emit"),
-        exitCode: 0,
-      },
-    ]);
-  });
-
-  it("returns a parse error when required command args are missing", async () => {
-    const commandCallsResult: BashAutomationCommandResult[] = [];
-    const context = createTestCommandContext();
-    const handlers: TestAutomationCommandHandlers = {
-      "otp.identity.create-claim": async () => {
-        return {
-          data: {
-            url: "https://example.com/claims",
-            externalId: "should-not-call",
-            code: "000000",
-          },
-        };
-      },
-      "automations.identity.lookup-binding": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-        },
-      }),
-      "automations.identity.bind-actor": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-          key: "external-id",
-          status: "linked",
-        },
-      }),
-      "event.emit": async () => ({
-        data: {
-          accepted: true,
-          eventId: "help-event",
-          source: "telegram",
-          eventType: "help-event",
-        },
-      }),
-      "scripts.run": async () => {
-        throw new Error("not available in test");
-      },
-    };
-
-    const bash = new Bash({
-      customCommands: createAutomationCommands(
-        TEST_COMMAND_SPEC_LIST,
-        handlers,
-        context,
-        commandCallsResult,
-      ),
-    });
-
-    const result = await bash.exec("otp.identity.create-claim --external-actor-id actor_1");
-
-    expect(result.exitCode).toBe(1);
-    expect(result.stderr).toContain("Missing required option --source");
-    expect(context.invocation).toEqual([]);
-    expect(commandCallsResult).toEqual([
-      {
-        command: "otp.identity.create-claim",
-        output: "",
-        exitCode: 1,
-      },
-    ]);
-  });
-
-  it("returns a parse error when ttl-minutes is not a positive integer", async () => {
-    const commandCallsResult: BashAutomationCommandResult[] = [];
-    const context = createTestCommandContext();
-    const handlers: TestAutomationCommandHandlers = {
-      "otp.identity.create-claim": async () => {
-        return {
-          data: {
-            url: "https://example.com/claims",
-            externalId: "should-not-call",
-            code: "000000",
-          },
-        };
-      },
-      "automations.identity.lookup-binding": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-        },
-      }),
-      "automations.identity.bind-actor": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-          key: "external-id",
-          status: "linked",
-        },
-      }),
-      "event.emit": async () => ({
-        data: {
-          accepted: true,
-          eventId: "help-event",
-          source: "telegram",
-          eventType: "help-event",
-        },
-      }),
-      "scripts.run": async () => {
-        throw new Error("not available in test");
-      },
-    };
-
-    const bash = new Bash({
-      customCommands: createAutomationCommands(
-        TEST_COMMAND_SPEC_LIST,
-        handlers,
-        context,
-        commandCallsResult,
-      ),
-    });
-
-    const result = await bash.exec(
-      "otp.identity.create-claim --source telegram --external-actor-id actor_1 --ttl-minutes 0",
-    );
-
-    expect(result.exitCode).toBe(1);
-    expect(result.stderr).toContain("--ttl-minutes must be a positive integer");
-    expect(context.invocation).toEqual([]);
-    expect(commandCallsResult).toEqual([
-      {
-        command: "otp.identity.create-claim",
-        output: "",
-        exitCode: 1,
-      },
-    ]);
-  });
-
-  it("omits pi.session commands from automation bash hosts", async () => {
-    const context = createTestCommandContext();
-    const { bash, commandCallsResult } = createBashHost({
-      fs: new InMemoryFs(),
-      context: {
-        automation: {
-          ...context,
-          runtime: {
-            emitEvent: runtime.emitEvent,
-          },
-        },
-        automations: {
-          runtime: {
-            lookupBinding: runtime.lookupBinding,
-            bindActor: runtime.bindActor,
-          },
-        },
-        otp: {
-          runtime: {
-            createClaim: runtime.createClaim,
-          },
-        },
-        pi: null,
-        reson8: null,
-        resend: null,
-        telegram: null,
-      },
-    });
-
-    const result = await bash.exec("pi.session.get --session-id session-1");
-
-    expect(result.exitCode).toBe(127);
-    expect(result.stderr).toContain("bash: pi.session.get: command not found");
-    expect(commandCallsResult).toEqual([]);
-  });
-
-  it("does not expose scripts.run inside automation execution hosts", async () => {
-    const event: AutomationEvent = {
-      id: "event-no-scripts-run",
-      orgId: "org-1",
-      source: "telegram",
-      eventType: "message.received",
-      occurredAt: "2026-01-01T00:00:00.000Z",
-      payload: {},
-    };
-
-    const automationRuntime = createStorageBackedAutomationRuntime({
-      hookContext: {
-        handlerTx: (() => {
-          throw new Error("handlerTx should not be used in this test");
-        }) as never,
-      },
-      event,
-    });
-
-    const { bash, commandCallsResult } = createBashHost({
-      fs: new InMemoryFs(),
-      context: createRuntimeAutomationExecutionContext({
-        event,
-        binding: {
-          source: "telegram",
-          eventType: "message.received",
-          scriptId: "script-no-scripts-run",
-        },
-        idempotencyKey: "idem-no-scripts-run",
-        runtime: automationRuntime,
-        pi: null,
-      }),
-    });
-
-    const result = await bash.exec(
-      "scripts.run --script /automations/test.sh --event /events/2026-03-26/evt-1.json",
-    );
-
-    expect(result.exitCode).toBe(1);
-    expect(result.stderr).toContain("scripts.run is not available in this execution context");
-    expect(commandCallsResult).toEqual([
-      {
-        command: "scripts.run",
-        output: "",
-        exitCode: 1,
-      },
-    ]);
-  });
-
-  it("passes shared automation command context through handler execution", async () => {
-    const commandCallsResult: BashAutomationCommandResult[] = [];
-    const context = createTestCommandContext({
-      idempotencyKey: "idem-shared",
-      orgId: "org-shared",
-      event: {
-        id: "event-shared",
-        source: "telegram",
-        eventType: "message.received",
-        occurredAt: "2026-01-01T00:00:00.000Z",
-        orgId: "org-shared",
-        payload: {
-          subject: "context-test",
-        },
-      },
-      binding: {
-        source: "telegram",
-        eventType: "message.received",
-        scriptId: "script-shared",
-      },
-      bashEnv: {},
-    });
-
-    let observedContext: {
-      eventId: string;
-      orgId?: string;
-      idempotencyKey: string;
-      binding: string;
-      scriptId: string;
-    } | null = null;
-
-    const handlers: TestAutomationCommandHandlers = {
-      "otp.identity.create-claim": async (_, commandContext) => {
-        observedContext = {
-          eventId: commandContext.event.id,
-          orgId: commandContext.orgId,
-          idempotencyKey: commandContext.idempotencyKey,
-          binding: commandContext.binding.source,
-          scriptId: commandContext.binding.scriptId,
-        };
-
-        return {
-          data: {
-            url: `https://example.com/${commandContext.binding.source}`,
-            externalId: "external-id",
-            code: "123456",
-          },
-        };
-      },
-      "automations.identity.lookup-binding": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-        },
-      }),
-      "automations.identity.bind-actor": async () => ({
-        data: {
-          value: "help-user",
-          source: "telegram",
-          key: "external-id",
-          status: "linked",
-        },
-      }),
-      "event.emit": async () => ({
-        data: {
-          accepted: true,
-          eventId: "event-ctx",
-          source: "telegram",
-          eventType: "message.received",
-        },
-      }),
-      "scripts.run": async () => {
-        throw new Error("not available in test");
-      },
-    };
-
-    const bash = new Bash({
-      customCommands: createAutomationCommands(
-        TEST_COMMAND_SPEC_LIST,
-        handlers,
-        context,
-        commandCallsResult,
-      ),
-    });
-
-    const result = await bash.exec(
-      "otp.identity.create-claim --source telegram --external-actor-id external-id",
-    );
-
-    expect(result.exitCode).toBe(0);
-    expect(observedContext).toEqual({
-      eventId: "event-shared",
-      orgId: "org-shared",
-      idempotencyKey: "idem-shared",
-      binding: "telegram",
-      scriptId: "script-shared",
-    });
-    expect(commandCallsResult).toEqual([
-      {
-        command: "otp.identity.create-claim",
-        output: "",
-        exitCode: 0,
-      },
-    ]);
-  });
-
   it("forwards sub-script stdout through scripts.run", async () => {
     const { bash, commandCallsResult } = createBashHost({
       fs: new InMemoryFs(),
